+++ NaturalnieApp/NaturalnieApp/Forms/ProductMenu/ShowProductInfo.cs
@@ -800,7 +800,7 @@ namespace NaturalnieApp.Forms
             try
             {
                 Validation.MariginValueValidation(localSender.Text);
-                this.ProductEntity.Marigin = float.Parse(localSender.Text);
+                this.ProductEntity.Marigin = Int32.Parse(localSender.Text);
                 //Update final price
                 UpdateFinalPrice();
                 errorProvider1.Clear();
